# Peeking lexer optimizations