Hi @alecthomas,

when rebasing #213 to master, I decided it would be best to commit my stuff in smaller chunks. The first chunk is some optimizations I did to `PeekingLexer` as it was becoming the bottleneck to the generated parser implementation.

1. **Optimizing `PeekingLexer.Clone` - `Checkpoint`**
    It does an allocation that could be avoided. One way is to just have `parseContext` and my generated code to store it by value (keeping pointer receiver for its methods though). This already helps quite a lot, but I went a bit further in my generated code - wrapping the internal cursors in a public `lexer.Checkpoint` type. When a branch needs a backup of the `PeekingLexer` state, it will just need to copy this small struct and restore it only if the branch was *not* accepted and it's even a little bit faster than copying & replacing the entire `PeekingLexer` struct.
2. **Optimizing `PeekingLexer.Clone` - removing it**
    Using the `Checkpoint` pattern in the reflection parser is tough as it needs to explicitly apply an accepted branch instead of reverting to a backup of a failed branch (the way that was nicer to use in the generated code). So my proposal would be to keep the `Checkpoint` pattern only for generated code. But I added a commit (independent of my other changes) that I can include if you like that stores `PeekingLexer` in `parseContext` by value and avoids the `Clone` allocation altogether.
3. **Optimizing `PeekingLexer.Peek`**
    Currently `Peek` (which is used quite often by the reflective parser and even more often by my generated parser) needs to skip elided tokens every time it's called. It also checks for the length of `PeekingLexer.tokens` and returns `p.eof` in a lot of places - I think it's more elegant to just have the EOF token as the last token and prevent `Next` from advancing beyond it.
    This optimization does this work once in `Upgrade` and `Next` and makes `Peek` super simple, it also removes the `eof` field. The performance improvement for the Thrift benchmark is barely noticeable, but a separate benchmark I added shows a >3x improvement for `Peek` calls; it was also more more noticeable in my generated parser.
    I mentioned this whole thing in point 2 in https://github.com/alecthomas/participle/issues/213#issuecomment-1055395492 as the initial implementation changed the behavior of `RawCursor` and thus also `Range`. This new implementation adds a separate cursor for knowing when the next non-elided token is, making it a bit slower, but prevents a change of behavior.

4. **Working with `lexer.Token` as pointer**
    `Token` is not a small struct and in most cases it's more efficient to work with it as a pointer instead of copying it - assuming of course no allocations are needed as the tokens are already stored in `PeekingLexer.tokens`. Changing `Token` methods to a pointer receiver alone improved CPU perf in the Thrift benchmark by ~2.5%, trying to make `Next`, `Peek` and other methods return it as a pointer made the improvement ~4%. Such a
    Considering this would be a breaking change to a public (somewhat internal, but still public) API, I decided it wasn't worth it. But if you think it's OK, I still added a new method `Current`, which is a version of `Peek` that returns a pointer. `BenchmarkPeekingLexer_Current` still shows that almost 3x faster than `Peek`. If you don't like it, I don't need to include it. I also used pointers at least inside the `PeekingLexer` methods and got a tiny CPU improvement in the Thrift benchmark out of it.

5. **`BatchLexer` interface**
    This is actually unrelated to the code generation, but when playing with a manually written lexer for Jinja, I found it both more convenient and more efficient (thanks to a much lower number of function calls and more efficient allocations) to have the ability to yield multiple (in my case up to 50) tokens at a time. The interface and the related implementation is just a soft suggestion, if you don't like it, I can remove it. But it's optional - if a `Lexer` chooses to implement it, it can gain extra performance, if it doesn't, no problem.

After all but the last optimization (`BatchLexer` isn't used by any built-in lexer - yet), the final comparison looks like:
```
BenchmarkParticipleThriftGenerated-Before-8   	   12105	    292277 ns/op	        27.46 MiB/s	  178816 B/op	    2167 allocs/op
BenchmarkParticipleThriftGenerated-After-8   	   12597	    269844 ns/op	        28.58 MiB/s	  161920 B/op	    1903 allocs/op
```

Perhaps even more important is that according to a profiler, now it spends almost 7x less time in the `PeekingLexer` methods which I intended to optimize.