@alecthomas 2022-08-13
Welcome back :)

These don't seem unreasonable, but do you have any benchmark comparisons if just lexing (vs. lexing+parsing)?