@petee-d 2022-11-07
Regarding 1., I just realized I didn't include `MakeCheckpoint` and `LoadCheckpoint` in this PR. Well, basically there are these 3 options:

1. Copying the entire lexer, would look like: `branchLexer := *p.Lex` and `*p.Lex = branchLexer`
2. `Checkpoint` struct manipulated using methods: `branchCheckpoint := p.Lex.MakeCheckpoint()` and `p.Lex.LoadCheckpoint(branchCheckpoint)`
3. `Checkpoint` struct used directly to save/load a checkpoint: `branchCheckpoint := p.Lex.Checkpoint` and `p.Lex.Checkpoint = branchCheckpoint`

I think when I started rebasing my changes (which initially used approach 2.), I decided to instead go for approach 3., but kinda forgot about it. Can't say I care too much about which of these approaches we ultimately pick. I lean towards 3. at the moment, I guess, seems the most elegant, plus has that tiny performance improvement of not copying immutable PeekingLexer fields?

By the way, I finally found time for coding today and I finished resolving conflicts against the current participle master - I gave up on using git merge (the conflicts were just too massive and difficult to read) and instead manually copy-pasted the changes from the original branch diff - was actually faster that way. :D My test suite for the Jinja template language is now passing.

I'm not even close to done though, as I still need to implement support for the 2 newly added node types (`custom` and `union`), see if there were any changes to participle behavior I need to reflect in my generated code, work my way through tens of TODOs I left for fixing some obscure differences from native parser behavior (especially error reporting and lookahead) and then actually make generating the parser reasonably easy. I will also need your input in https://github.com/alecthomas/participle/issues/264.