@petee-d 2022-11-08
@klondikedragon I ran a benchmark (parsing only on Jinja) and the Checkpoint optimization was actually even more impactful than I remembered. Benchmarks were run with `-test.count=3 -test.benchtime=3s`, picking the best of 3, each line starts with the approach number in my post above.

```
1: BenchmarkParseOnly-8   	   24804	    145576 ns/op	   99288 B/op	     792 allocs/op
2: BenchmarkParseOnly-8   	   26631	    135601 ns/op	   99288 B/op	     792 allocs/op
3: BenchmarkParseOnly-8   	   25071	    135069 ns/op	   99288 B/op	     792 allocs/op
```

Unsurprisingly, 2. and 3. were equal as the `MakeCheckpoint` and `LoadCheckpoint` methods are inlined anyway. Comparing just copying the entire lexer vs Checkpoint:
```
BenchmarkPeekingLexer_CopyWhole-8   		829165962	         1.421 ns/op
BenchmarkPeekingLexer_CopyCheckpoint-8   	1000000000	         0.3934 ns/op
```

I got curious and tried adding an unused map or a slice to Checkpoint, the slice actually slowed it down the most, noticeably more than adding 24 bytes of non-pointer data. Not entirely sure what goes on on the background, but maybe Go has to be more careful manipulating pointer data due to garbage collection? (there probably is a reason why there is a separate `memclrNoHeapPointers` in the runtime)

Anyway, I wouldn't say performance is the main reason why I prefer approach 3. - I just find it more readable.