@klondikedragon 2022-11-09
@petee-d -- pretty interesting the copy of the slice was the biggest source of slowdown in copying the struct, I wouldn't have guessed that! It makes me curious to learn more about how Go does GC. Apparently there is a build flag where it will output escape analysis info, `go build -gcflags="-m" ...`

@alecthomas - @petee-d makes a compelling case for the "Checkpoint" style modification to `PeekingLexer`, with a ~7% overall speedup of parsing, and the "3" approach described above is copying/restoring the Checkpoint field of the struct instead of the whole struct itself, so pretty similar semantics to the current code. What do you think?